Ball Shooting Finished
The balls now shoot from the gun! They shoot out with a bit of force in the direction the player is facing, fall in the opposite direction of the player, and change gravity when the player does. Idk if y'all wanted an unlimited number or if we wanted the ball to despawn every time a new one is shot, but let me know if it's the latter because I think I know how to do that but right now you can just shoot indefinitely